==> 02_Calidad de Datos.ipynb <==
# CALIDAD DE DATOS

## IMPORTAR PAQUETES

import numpy as np
import pandas as pd
from janitor import clean_names

#Automcompletar rápido
%config IPCompleter.greedy=True

## IMPORTAR LOS DATOS

Sustituir la ruta del proyecto.

ruta_proyecto = 'C:/Users/Ana/Desktop/DATA_SCIENCE_FOR_BUSINESS/PYTHON_DATA_SCIENCE_MASTERY/EstructuraDirectorio/03_MACHINE_LEARNING/08_CASOS_/01_LEADSCORING'

Sustituir el nombre_fichero_datos.

nombre_fichero_datos = 'trabajo.csv'

#En este caso no trabajamos sobre una muestra, sino directamente sobre el dataset de trabajo

Cargar los datos.

ruta_completa = ruta_proyecto + '/02_Datos/03_Trabajo/' + nombre_fichero_datos

df = pd.read_csv(ruta_completa,index_col='id', sep = ',')
df

## CORRECCIÓN DE NOMBRES DE VARIABLES

df.info()

#Los nombres estan bien, por lo que no es necesario realizar ningun tipo de corrección
#Elimino los pasos de Corrección de este Notebook

## VISIÓN GENERAL

df.head(5)

df.info()

## TIPOS DE DATOS

### Identificación

df.dtypes

### Corrección

df = df.astype({'visitas_total':'Int64'})

df.dtypes

## VALORES ÚNICOS

### Identificación

df.conociste_youtube.value_counts(normalize = True)

df.nunique().sort_values()

#Las variables que tienen un único valor son Constantes, por lo que hay que eliminarlas
#No hay variables con excesivo número de valores

### Corrección

df.drop(columns = ['conociste_revista','conociste_youtube'], inplace = True)

df.nunique().sort_values()

## DUPLICADOS

### Identificación

df.duplicated().sum()

### Corrección

df.drop_duplicates(inplace = True)

df.info()

## SEPARAR NUMÉRICAS Y CATEGÓRICAS

### Categóricas

cat = df.select_dtypes(exclude = 'number').copy()

cat

### Numéricas

num = df.select_dtypes(include='number').copy()

num

## GESTIÓN DE CATEGÓRICAS

### Nulos

#### Identificación

cat.isna().sum().sort_values(ascending = False)

#ocupación no la podemos eliminar ya que es una variable importante en el leadscoring 
#No vamos a eliminar ninguna Variable ni ningún Registro

cat.shape

num.shape

#### Corrección

cat.ocupacion.value_counts()

#Hay una diferencia muy alta entre desempleado y el resto de categorias, por ello vamos a Imputar ocupacion por la Moda

cat.ambito.value_counts()

#Imputamos select (información no dada por el usuario) por la Moda

cat.ult_actividad.value_counts()

cat.fuente.value_counts()

##### Eliminar la variable

##### Eliminar registros

##### Imputar por un valor

###### Variables a imputar

cat.columns.to_list()

var_imputar_valor = ['origen',
                     'fuente',
                     'no_enviar_email',
                     'no_llamar',
                     'ult_actividad',
                     'conociste_google',
                     'conociste_periodico',
                     'conociste_facebook',
                     'conociste_referencias',
                     'descarga_lm']


###### Valor por el que imputar

valor = 'NULO'

###### Imputar

cat[var_imputar_valor] = cat[var_imputar_valor].fillna(valor)

##### Imputar por la moda

###### Variables a imputar

var_imputar_moda = ['ocupacion','ambito']

###### Funcion para imputar con la moda

def imputar_moda(variable):
    return(variable.fillna(variable.mode()[0]))

###### Imputar con la moda

cat[var_imputar_moda] = cat[var_imputar_moda].apply(imputar_moda)

cat.isna().sum().sort_values(ascending = False)

### Atípicos

#### Categorías poco frecuentes

##### Función para agrupar categorías raras en 'OTROS'

def agrupar_cat_raras(variable, criterio = 0.02):
    #Calcula las frecuencias
    frecuencias = variable.value_counts(normalize=True)
    #Identifica las que están por debajo del criterio
    temp = [cada for cada in frecuencias.loc[frecuencias < criterio].index.values]
    #Las recodifica en 'OTROS'
    temp2 = np.where(variable.isin(temp),'OTROS',variable)
    #Devuelve el resultado
    return(temp2)

##### Variables a imputar

cat.columns.to_list()

var_imputar_moda = cat.columns.to_list()

#Imputamos todas las variables con pocos valores 

##### Aplicar

6000 * 0.02

Se imputarán todas aquellas variables que tengan un conteo inferior a 120 registos

for variable in var_imputar_moda:
    cat[variable] = agrupar_cat_raras(cat[variable],criterio = 0.02)

cat.ocupacion.value_counts()

cat.ambito.value_counts()

cat.ult_actividad.value_counts()

cat.fuente.value_counts()

Ya no tenemos Atípicos en las Variables Categóricas

## GESTIÓN DE NUMÉRICAS

### Nulos

#### Identificación

num

num.isna().sum().sort_values(ascending = False)

#### Corrección

##### Imputar por la mediana

###### Variables a imputar

var_imputar_mediana = num.columns.to_list()
num.columns.to_list()

###### Funcion para imputar con la mediana

def imputar_mediana(variable):
    if pd.api.types.is_integer_dtype(variable):
        return(variable.fillna(int(variable.median())))
    else:
        return(variable.fillna(variable.median()))

###### Imputar con la mediana

num[var_imputar_mediana] = num[var_imputar_mediana].apply(imputar_mediana)

num.isna().sum().sort_values(ascending = False)

### Atípicos

#### Por desviación típica

##### Identificación

###### Definir el número de desviaciones típicas

num_desv_tip = 4

###### Función que devuelve los índices de los atípicos

def atipicos_desv_tip(variable, num_desv_tip = 4):
    #sacamos los nulos por ahora
    variable = variable.dropna()
    #calculamos los límites
    media = np.mean(variable)
    sd = np.std(variable)
    umbral = sd * num_desv_tip
    lim_inf = media - umbral
    lim_sup = media + umbral
    #encontramos los índices de los que están fuera de los límites
    indices = [indice for indice,valor in variable.items() if valor < lim_inf or valor > lim_sup]
    return(indices)

###### Función que cuenta el número de atípicos

def conteo_atipicos(df,variable, num_desv_tip = 4):
    atipicos = atipicos_desv_tip(df[variable], num_desv_tip)
    return(df.loc[atipicos,variable].value_counts())

###### Variables a aplicar

var_atipicos_dt = num.columns.to_list()

###### Aplicar

for variable in var_atipicos_dt:
    print('\n' + variable + ':\n')
    print(conteo_atipicos(num,variable,num_desv_tip))

num.info()

###### Aplicar

num['visitas_total'].clip(0,50, inplace=True)
num['paginas_vistas_visita'].clip(0,20, inplace=True)

## GUARDAR DATASETS TRAS CALIDAD DE DATOS

Guarda los avances en df, cat y num con un sufijo para poder guardar o recuperar avances del proyecto.

En formato pickle para no perder las modificaciones de metadatos.

### Definir los nombres de los archivos

ruta_trabajo = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'trabajo_resultado_calidad.pickle'
ruta_cat = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'cat_resultado_calidad.pickle'
ruta_num = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'num_resultado_calidad.pickle'

### Guardar los archivos

df.to_pickle(ruta_trabajo)
cat.to_pickle(ruta_cat)
num.to_pickle(ruta_num)

df.info()
