==> 04_Transformacion de datos.ipynb <==
# TRANSFORMACIÓN DE DATOS

## IMPORTAR PAQUETES

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import OrdinalEncoder
from sklearn.preprocessing import LabelEncoder
from category_encoders import TargetEncoder
from sklearn.preprocessing import KBinsDiscretizer
from sklearn.preprocessing import Binarizer
from sklearn.preprocessing import PowerTransformer
from sklearn.preprocessing import QuantileTransformer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import RobustScaler
from sklearn.preprocessing import MaxAbsScaler
from imblearn.under_sampling import RandomUnderSampler
from imblearn.over_sampling import RandomOverSampler

#Automcompletar rápido
%config IPCompleter.greedy=True

## IMPORTAR LOS DATOS

1.- Sustituir la ruta del proyecto.

ruta_proyecto ='C:/Users/Ana/Desktop/DATA_SCIENCE_FOR_BUSINESS/PYTHON_DATA_SCIENCE_MASTERY/EstructuraDirectorio/03_MACHINE_LEARNING/08_CASOS_/01_LEADSCORING'

2.- Nombrar los ficheros de datos.

nombre_cat = 'cat_resultado_eda.pickle'
nombre_num = 'num_resultado_eda.pickle'

3.- Cargar los datos.

cat = pd.read_pickle(ruta_proyecto + '/02_Datos/03_Trabajo/' + nombre_cat).reset_index(drop = True)
num = pd.read_pickle(ruta_proyecto + '/02_Datos/03_Trabajo/' + nombre_num).reset_index(drop = True)

4.- Separar la target.

target = num[['compra']].copy().reset_index(drop=True)

## TRANSFORMACIÓN DE CATEGÓRICAS

### One Hot Encoding

#### Variables a aplicar OHE

cat.info()

var_ohe = ['origen','fuente','ult_actividad','ambito', 'ocupacion','descarga_lm']

#### Instanciar

ohe = OneHotEncoder(sparse = False, handle_unknown='ignore')

#### Entrenar y aplicar

cat_ohe = ohe.fit_transform(cat[var_ohe])

#### Guardar como dataframe

cat_ohe = pd.DataFrame(cat_ohe, columns = ohe.get_feature_names_out())

## UNIFICAR DATASETS TRANSFORMADOS

### Unir todos los dataframes

df = pd.concat([cat_ohe, num], axis = 1)
df

## REESCALAR VARIABLES

### Con Min-Max

df.info()

#### Variables a reescalar con Min-Max

var_mms = df.iloc[:,39:].columns

var_mms

#### Instanciar

mms = MinMaxScaler()

#### Entrenar y aplicar

df_mms = mms.fit_transform(df[var_mms])

#### Guardar como dataframe

#Añadir sufijos a los nombres
nombres_mms = [variable + '_mms' for variable in var_mms]

#Guardar como dataframe
df_mms = pd.DataFrame(df_mms,columns = nombres_mms)

## UNIFICAR DATASETS REESCALADOS

id = df.id

pd.concat([id,cat_ohe, df_mms, target], axis = 1).set_index('id') #ponemos id como indice 

#Comprobamos que son todo numéricas y en rango 0/1
pd.concat([id,cat_ohe, df_mms, target], axis = 1).set_index('id').describe().T

### Unir todos los dataframes en el tablón analítico

df_tablon = pd.concat([id,cat_ohe, df_mms, target], axis = 1).set_index('id')

df_tablon

## GUARDAR DATASET TRAS TRANSFORMACIÓN DE DATOS

En formato pickle para no perder las modificaciones de metadatos.

#Definir los nombres del archivo
ruta_df_tablon = ruta_proyecto + '/02_Datos/03_Trabajo/' + 'df_tablon.pickle'

#Guardar los archivos
df_tablon.to_pickle(ruta_df_tablon)
