==> 08_Preparacion del codigo de produccion.ipynb <==
# **Preparacion del codigo de produccion**

## IMPORTAR LAS LIBRERIAS

Actualizar a las que se usen finalmente en tu proyecto.

import numpy as np
import pandas as pd
import cloudpickle

#Automcompletar rápido
%config IPCompleter.greedy=True

from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import MinMaxScaler

from sklearn.linear_model import LogisticRegression

from sklearn.pipeline import Pipeline
from sklearn.preprocessing import FunctionTransformer
from sklearn.compose import make_column_transformer
from sklearn.pipeline import make_pipeline

## CARGAR LOS DATOS

### Ruta del proyecto

ruta_proyecto =  'C:/Users/Ana/Desktop/DATA_SCIENCE_FOR_BUSINESS/PYTHON_DATA_SCIENCE_MASTERY/EstructuraDirectorio/03_MACHINE_LEARNING/08_CASOS_/01_LEADSCORING'

### Nombre del fichero de datos

nombre_fichero_datos = 'Leads.csv'

### Cargar los datos

ruta_completa = ruta_proyecto + '/02_Datos/01_Originales/' + nombre_fichero_datos

df = pd.read_csv(ruta_completa,index_col='id', sep = ';') # Hay que coger los datos de la carga de datos del setup

df

### Seleccionar solo las variables finales

#### Cargar la lista de variables finales

nombre_variables_finales = ruta_proyecto + '/05_Resultados/' + 'variables_finales.pickle'

pd.read_pickle(nombre_variables_finales).sort_index().index.to_list()

#### Apuntar (manualmente) la lista de variables finales sin extensiones

variables_finales = ['ambito',
                     'descarga_lm',
                     'ocupacion',
                     'paginas_vistas_visita',
                     'score_actividad',
                     'score_perfil', 'tiempo_en_site_total', 
                     'ult_actividad',
                     'visitas_total'
                  ]

#### Crear la matriz de variables procesos (excel)

Ir a la plantilla de Excel "Fase Producción Plantilla Procesos" y crear la matriz de variables por procesos.

#### Actualizar las importaciones

Ir arriba a la celda de importacion de paquetes y actualizarlos con los que finalmente vamos a usar.

## ESTRUCTURA DE LOS DATASETS

df.shape

### Eliminar registros

#### Por duplicados

df.drop_duplicates(inplace = True)

df.shape

#### Por EDA

df = df.loc[(df.no_enviar_email != 'Yes') & (df.no_llamar != 'OTROS') & (df.ult_actividad != 'Email Bounced')]

df.shape

#### Para x

Quedarse solo con las de la lista.

x = df[variables_finales].copy()

x

#### Para y

Especificar la target.

target = 'compra'

Crear el y.

y = df[target].copy()

y

## CREAR EL PIPELINE

### Instanciar calidad de datos

#### Crear la función

x.info()

def calidad_datos(df):
    
    temp = df.astype({'visitas_total': 'Int64'}) 
    
    def imputar_moda(variable):
        return(variable.fillna(variable.mode()[0]))
    
    var_imputar_moda = ['ocupacion','ambito']
    
    temp[var_imputar_moda] = temp[var_imputar_moda].apply(imputar_moda)
    
    var_imputar_valor = ['descarga_lm',
                     'ult_actividad']
    
    valor = 'NULO'
    
    temp[var_imputar_valor] = temp[var_imputar_valor].fillna(valor)
    
    var_imputar_mediana = ['paginas_vistas_visita',
                           'score_actividad',
                           'score_perfil', 
                           'tiempo_en_site_total', 
                           'visitas_total']
    
    def imputar_mediana(variable):
        if pd.api.types.is_integer_dtype(variable):
            return(variable.fillna(int(variable.median())))
        else:
            return(variable.fillna(variable.median()))
        
    temp[var_imputar_mediana] = temp[var_imputar_mediana].apply(imputar_mediana)
    
    def agrupar_cat_raras(variable, criterio = 0.02):
        #Calcula las frecuencias
        frecuencias = variable.value_counts(normalize=True)
        #Identifica las que están por debajo del criterio
        temp = [cada for cada in frecuencias.loc[frecuencias < criterio].index.values]
        #Las recodifica en 'OTROS'
        temp2 = np.where(variable.isin(temp),'OTROS',variable)
        #Devuelve el resultado
        return(temp2)
    
    var_agrupar_cat_raras = ['ocupacion',
                             'ambito',
                             'descarga_lm',
                             'ult_actividad']
    
    for variable in var_agrupar_cat_raras:
        temp[variable] = agrupar_cat_raras(temp[variable],criterio = 0.02)
    
    
    
    temp['visitas_total'].clip(0,50, inplace=True) # Winsorizacion manual
    
    return(temp)

x.shape

calidad_datos(x).isna().sum()

#### Convertirla en transformer

hacer_calidad_datos = FunctionTransformer(calidad_datos)

### Instanciar transformación de variables

var_ohe = ['ambito', 'descarga_lm', 'ocupacion', 'ult_actividad']
ohe = OneHotEncoder(sparse = False, handle_unknown='ignore')



var_mms = ['paginas_vistas_visita','score_actividad','score_perfil','tiempo_en_site_total', 'visitas_total']
mms = MinMaxScaler()

### Crear el pipe del preprocesamiento

#### Crear el column transformer

ct = make_column_transformer(
    (ohe, var_ohe),
    (mms, var_mms),
    remainder='drop')

#### Crear el pipeline del preprocesamiento

pipe_prepro = make_pipeline(hacer_calidad_datos, # Pandas
                            ct) #Scikitlearn

### Instanciar el modelo

#### Instanciar el algoritmo

modelo = LogisticRegression(n_jobs = -1, 
                       solver = 'saga',
                       penalty = 'none',
                       C = 0.25 # podriamos quitar C al tener penalty = 'none'
                       )

#### Crear el pipe final de entrenamiento

pipe_entrenamiento = make_pipeline(pipe_prepro,modelo)

#### Guardar el pipe final de entrenamiento

nombre_pipe_entrenamiento = 'pipe_entrenamiento.pickle'

ruta_pipe_entrenamiento = ruta_proyecto + '/04_Modelos/' + nombre_pipe_entrenamiento

with open(ruta_pipe_entrenamiento, mode='wb') as file:
   cloudpickle.dump(pipe_entrenamiento, file)

#### Entrenar el pipe final de ejecución

pipe_ejecucion = pipe_entrenamiento.fit(x,y)

## GUARDAR EL PIPE

### Nombre del pipe final de ejecución

nombre_pipe_ejecucion = 'pipe_ejecucion.pickle'

### Guardar el pipe final de ejecución

ruta_pipe_ejecucion = ruta_proyecto + '/04_Modelos/' + nombre_pipe_ejecucion

with open(ruta_pipe_ejecucion, mode='wb') as file:
   cloudpickle.dump(pipe_ejecucion, file)
